一覧と個別ページ両方で使うのでキャラクター情報をカスタムフックに切り出す
